Added two more buttons, animations, and a server/WASM indicator
Added two more buttons "Create Match" and "List Matches". Also renamed
"Find Match" to "Quick Match"

Added animations that move the clicked button to the upper left
corner and hides the others.

Added an indicator the the bottom right corner that says whether the app
is running on the server or in the browser using WASM.